Murders in New Zealand (2004 - 2019)

This dataset contains information about recorded murder cases in New Zealand between 2004 and 2019. It includes details on the sex, age, and cause of death of the victims, as well as the identity of the alleged killer, the date of the crime, and the region where the crime occurred. The dataset is in the form of a simple features (sf) object, with geographic data represented as points.

Usage

data(nz_murders_sf)

Format

An sf data frame with 967 observations and 12 variables:

sex

Sex of the victim (character).

age

Age of the victim (integer).

date

Date of the murder (character).

year

Year the murder occurred (integer).

cause

Cause of death (character).

killer

Name of the alleged killer (character).

name

Name of the victim (character).

full_date

Full date and time of the murder (POSIXct).

month

Month of the murder (ordered factor with 12 levels).

cause_cat

Category of the cause of death (character).

region

Region where the murder occurred (character).

geometry

Geographic coordinates (sf POINT) representing the location of the murder (list of 967).

Source

Recorded murder data for New Zealand (2004 - 2019).

Details

The dataset name has been changed to 'nz_murders_sf' to avoid confusion with other datasets in the R ecosystem. This naming convention helps distinguish this dataset as part of the crimedatasets package and assists users in identifying its specific characteristics. The suffix '_sf' indicates that the dataset is an sf object in R, used for storing and handling spatial data. The original content has not been modified in any way.